+++ e2e/src/utils/mod.rs
@@ -5,6 +5,7 @@ pub mod hyperlane;
 pub mod ibc;
 pub mod icq;
 pub mod manager;
+pub mod mock_cctp_relayer;
 pub mod osmosis;
 pub mod parse;
 pub mod persistence;
